+++ defaults.py
@@ -639,8 +639,9 @@ class FlatCAMDefaults:
         # Markers Tool
         "tools_markers_thickness": 0.1,
         "tools_markers_length": 3.0,
-        "tools_markers_marginx": 0.0,
-        "tools_markers_marginy": 0.0,
+        "tools_markers_reference": "e",     # from edge
+        "tools_markers_offset_x": 0.0,
+        "tools_markers_offset_y": 0.0,
         "tools_markers_type": 's',
         "tools_markers_drill_dia": 0.5,